правки по заполнению тумб
